refactor with new component
